# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PakeLagi_API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
...
    "paths": {
...
        "/v1/admin/events/{event}": {
            "get": {
                "operationId": "events.show",
                "summary": "Detail event (admin)",
                "tags": [
                    "Admin - Event"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AdminEventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AdminEventResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "images"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "events.update",
                "summary": "Perbarui data event (admin)",
                "tags": [
                    "Admin - Event"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AdminEventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AdminEventResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "images"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "events.destroy",
                "summary": "Hapus event (admin)",
                "tags": [
                    "Admin - Event"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/admin/events/{event}/images/{image}": {
            "delete": {
                "operationId": "eventImage.destroy",
                "summary": "Hapus satu foto dokumentasi event",
                "tags": [
                    "Admin - Event Images"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/admin/products/{product}/images/{image}": {
            "patch": {
                "operationId": "productImage.update",
                "summary": "Atur foto utama produk",
                "tags": [
                    "Admin - Foto Produk"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductImageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AdminProductImageResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AdminProductImageResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "productImage.destroy",
                "summary": "Hapus foto produk",
                "tags": [
                    "Admin - Foto Produk"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "path",
                        "required": true,
                        "description": "The image ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/admin/categories/{category}": {
            "put": {
                "operationId": "categories.update",
                "summary": "Perbarui data kategori",
                "tags": [
                    "Admin - Kategori"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AdminCategoryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AdminCategoryResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "categories.destroy",
                "summary": "Hapus kategori",
                "tags": [
                    "Admin - Kategori"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "description": "The category ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/admin/products/{product}": {
            "get": {
                "operationId": "products.show",
                "summary": "Detail produk (admin)",
                "tags": [
                    "Admin - Produk"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AdminProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AdminProductResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "category",
                                                        "images"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "products.update",
                "summary": "Perbarui data produk",
                "tags": [
                    "Admin - Produk"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AdminProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AdminProductResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "category",
                                                        "images"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "products.destroy",
                "summary": "Hapus produk",
                "tags": [
                    "Admin - Produk"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminCategoryResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "measurement_fields": {
                        "type": [
                            "array",
                            "null"
                        ],
                        "items": {
                            "type": "string"
                        }
                    },
                    "products_count": {
                        "type": "integer"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "slug",
                    "measurement_fields"
                ],
                "title": "AdminCategoryResource"
            },
            "AdminEventResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "event_date": {
                        "type": "string"
                    },
                    "end_date": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "location": {
                        "type": "string"
                    },
                    "organizer": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "cover_image_url": {
                        "type": "null"
                    },
                    "cover_image_thumbnail_url": {
                        "type": "null"
                    },
                    "status": {
                        "type": "string"
                    },
                    "timing": {
                        "type": "string",
                        "enum": [
                            "upcoming",
                            "past"
                        ]
                    },
                    "is_upcoming": {
                        "type": "string"
                    },
                    "show_on_homepage": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EventImageResource"
                        }
                    }
                },
                "required": [
                    "id",
                    "slug",
                    "title",
                    "description",
                    "event_date",
                    "end_date",
                    "location",
                    "organizer",
                    "cover_image_url",
                    "cover_image_thumbnail_url",
                    "status",
                    "timing",
                    "is_upcoming",
                    "show_on_homepage",
                    "created_at",
                    "updated_at"
                ],
                "title": "AdminEventResource"
            },
            "AdminProductImageResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "url": {
                        "type": "string"
                    },
                    "thumbnail_url": {
                        "type": "null"
                    },
                    "is_primary": {
                        "type": "boolean"
                    },
                    "sort_order": {
                        "type": "integer"
                    }
                },
                "required": [
                    "id",
                    "url",
                    "thumbnail_url",
                    "is_primary",
                    "sort_order"
                ],
                "title": "AdminProductImageResource"
            },
...
            "EventImageResource": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "thumbnail_url": {
                        "type": "string"
                    },
                    "caption": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "sort_order": {
                        "type": "integer"
                    }
                },
                "required": [
                    "url",
                    "thumbnail_url",
                    "caption",
                    "sort_order"
                ],
                "title": "EventImageResource"
            },
...
            "SaveCategoryRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "measurement_fields": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "pattern": "^[a-z][a-z0-9_]*$",
                            "maxLength": 50
                        },
                        "maxItems": 12,
                        "uniqueItems": true
                    }
                },
                "required": [
                    "name",
                    "measurement_fields"
                ],
                "title": "SaveCategoryRequest"
            },
            "SaveEventRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 10000
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_date": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "organizer": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    },
                    "status": {
                        "$ref": "#/components/schemas/EventStatus"
                    },
                    "show_on_homepage": {
                        "type": "boolean"
                    },
                    "cover_image": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "binary",
                        "contentMediaType": "application/octet-stream",
                        "description": "Maximum file size: 2048 kilobytes."
                    },
                    "remove_cover": {
                        "type": [
                            "boolean",
                            "null"
                        ]
                    }
                },
                "required": [
                    "title",
                    "description",
                    "event_date",
                    "location",
                    "status"
                ],
                "title": "SaveEventRequest"
            },
...
            "UpdateProductImageRequest": {
                "type": "object",
                "properties": {
                    "is_primary": {
                        "enum": [
                            "yes",
                            "on",
                            "1",
                            1,
                            "true",
                            true
                        ]
                    }
                },
                "required": [
                    "is_primary"
                ],
                "title": "UpdateProductImageRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}